project: Switch development environment to uv
Requires Python 3.9 now.
