+++ .gitignore
@@ -7,6 +7,7 @@ __pycache__/
 dist/
 AUTHORS
 ChangeLog
+.uv-cache/
 
 # Tests
 .coverage
