+++ pyproject.toml
@@ -1,0 +1,48 @@
+[project]
+name = "wireshark_remote"
+description = "Initiate wireshark remote capture (SSH or AVM FRITZ!Box)"
+version = "0.1.0"
+authors = [
+  {name = "Christian Fetzer", email = "fetzer.ch@gmail.com"},
+]
+license = "MIT"
+readme = "README.md"
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Environment :: Console",
+    "Intended Audience :: System Administrators",
+    "Natural Language :: English",
+    "Operating System :: POSIX",
+    "Programming Language :: Python",
+    "Topic :: System :: Systems Administration",
+]
+keywords = [
+    "wireshark",
+    "ssh",
+]
+requires-python = ">=3.9"
+dependencies = [
+    "click>=8.1.8",
+]
+[dependency-groups]
+dev = [
+    "pylint>=3.3.7",
+    "pytest>=8.4.1",
+    "pytest-cov>=6.2.1",
+]
+
+[project.urls]
+Repository = "https://github.com/fetzerch/wireshark_remote"
+
+[project.scripts]
+wireshark-ssh = "wireshark_remote.cli:wireshark_ssh"
+wireshark-fritzbox = "wireshark_remote.cli:wireshark_fritzbox"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.coverage.run]
+omit = [
+    "wireshark_remote/fritzbox_web.py",
+]
