--- setup.cfg
@@ -1,26 +1,0 @@
-[metadata]
-name = wireshark_remote
-summary = Initiate wireshark remote capture (SSH or AVM FRITZ!Box)
-description-file = README.md
-description-content-type = text/markdown
-url = https://github.com/fetzerch/wireshark_remote
-author = Christian Fetzer
-author_email = fetzer.ch@gmail.com
-license = MIT
-classifiers =
-    Development Status :: 4 - Beta
-    Environment :: Console
-    Intended Audience :: System Administrators
-    License :: OSI Approved :: MIT License
-    Natural Language :: English
-    Operating System :: POSIX
-    Programming Language :: Python
-    Topic :: System :: Systems Administration
-
-[files]
-packages = wireshark_remote
-
-[entry_points]
-console_scripts =
-   wireshark-ssh = wireshark_remote:wireshark_ssh
-   wireshark-fritzbox = wireshark_remote:wireshark_fritzbox
