--- setup.py
@@ -1,5 +1,0 @@
-"""wireshark_remote - Initiate wireshark remote capture (SSH or AVM FRITZ!Box)"""
-
-from setuptools import setup
-
-setup(setup_requires=["pbr"], pbr=True)
