+++ tox.ini
@@ -1,16 +1,17 @@
+
+[gh]
+python =
+    3.9 = 3.9
+    3.10 = 3.10
+    3.11 = 3.11
+    3.12 = 3.12
+    3.13 = 3.13
+
 [tox]
-envlist = py37,py38,py39,py310,py311
+envlist = py39,py310,py311,py312,py313
 
 [testenv]
-deps = pytest-black
-       pytest-cov
-       pytest-mock
-       pytest-pylint
-commands = pytest {posargs} --black --pylint \
-                            --cov=wireshark_remote \
+deps = pytest-cov
+commands = pytest {posargs} --cov \
                             --cov-report=term-missing \
-                            --cov-branch --cov-fail-under=100
-
-[coverage:run]
-omit = .tox/*
-       wireshark_remote/fritzbox_web.py
+                            --cov-branch
